requirements.txt
torch>=2.0.1
safetensors==0.3.1
sentencepiece>=0.1.97
ninja==1.11.1
tiktoken==0.3.3
ninja
hvac
pynvml
psutil
pika
transformers
bitsandbytes
scipy
transformers-stream-generator
jsonschema
omegaconf
Pillow
einops
protobuf
accelerate
diffusers
timm
openai==0.28